% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTSNE.R
\name{WTSNE}
\alias{WTSNE}
\title{weigth t-Distributed Stochastic Neighbor Embedding}
\usage{
WTSNE(X, ...)
}
\arguments{
\item{X}{matrix; Data matrix (each row is an observation, each column is a variable)}

\item{...}{Other arguments that can be passed to Rtsne}
}
\value{
List with the following elements:
\item{Y}{Matrix containing the new representations for the objects}
\item{N}{Number of objects}
\item{origD}{Original Dimensionality before TSNE (only when \code{X} is a data matrix)}
\item{perplexity}{See above}
\item{theta}{See above}
\item{costs}{The cost for every object after the final iteration}
\item{itercosts}{The total costs (KL-divergence) for all objects in every 50th + the last iteration}
\item{stop_lying_iter}{Iteration after which the perplexities are no longer exaggerated}
\item{mom_switch_iter}{Iteration after which the final momentum is used}
\item{momentum}{Momentum used in the first part of the optimization}
\item{final_momentum}{Momentum used in the final part of the optimization}
\item{eta}{Learning rate}
\item{exaggeration_factor}{Exaggeration factor used to multiply the P matrix in the first part of the optimization}
}
\description{
weigth t-Distributed Stochastic Neighbor Embedding
}
